fix: deploy commands not using screenbuff
